+++ core/src/Sequence.h
@@ -20,28 +20,28 @@ public:
 
     T &GetLast() { return At(this->GetLength() - 1); }
 
-    void Set(int index, T value) {
+    void Set(size_t index, T value) {
         At(index) = value;
     }
 
-    virtual T &At(int index) = 0;
+    virtual T &At(size_t index) = 0;
 
-//    virtual int GetLength() = 0;
+//    virtual size_t GetLength() = 0;
 
-    T &operator[](unsigned int index) { return At(index); }
+    T &operator[](size_t index) { return At(index); }
 
     //Operations
     virtual void Append(T item) = 0;
 
     virtual void Prepend(T item) = 0;
 
-    virtual void InsertAt(T item, int index) = 0;
+    virtual void InsertAt(T item, size_t index) = 0;
 
     virtual void PopFirst() = 0;
 
     virtual void PopLast() = 0;
 
-    virtual void RemoveAt(int index) = 0;
+    virtual void RemoveAt(size_t index) = 0;
 
     virtual Sequence<T> *Concat(Sequence<T> &list) = 0;
